filter-repo: attempt to fix --prune-degenerate=never with path filtering
Gaurav Kanoongo reports that on a repository of his, that combining
`--prune-degenerate never` with a `--path` specification was hitting a
traceback, trying to write a parent out to the fast import stream whose
actual value was None.  I suspect that's because the parents of the
merge commit were filtered out by the path specification, leaving us
only with no-longer-extant parents.  Those need to be filtered out, we
just don't want to remove parents that are the same as or are an
ancestor of another parent.

Signed-off-by: Elijah Newren <newren@gmail.com>